Backend/requirements.txt
rich==13.6.0
pandas==1.5.3
numpy==1.25.2
psycopg2-binary==2.9.6
python-dotenv==1.0.0
keras
tensorflow
scikit-learn==1.6.1
openpyxl==3.1.5
xlrd==2.0.1
requests==2.32.3
numba==0.61.2
shap==0.47.2
matplotlib==3.10.1
lime==0.2.0.1
xgboost==3.0.0
dice-ml==0.11
ipython==8.35.0